improve useWatch with useSubscribe
