+++ src/useWatch.ts
@@ -1,5 +1,6 @@
 import * as React from 'react';
 
+import { generateWatchOutput } from './logic/generateWatchOutput';
 import shouldSubscribeByName from './logic/shouldSubscribeByName';
 import isObject from './utils/isObject';
 import isUndefined from './utils/isUndefined';
@@ -59,12 +60,10 @@ export function useWatch<TFieldValues>(props?: UseWatchProps<TFieldValues>) {
     subject: control._subjects.watch,
     callback: (formState) => {
       if (shouldSubscribeByName(_name.current, formState.name)) {
-        control._stateFlags.mount = true;
-        const fieldValues = control._getWatch(
-          _name.current as InternalFieldName,
-          defaultValue as UnpackNestedValue<
-            DeepPartialSkipArrayKey<TFieldValues>
-          >,
+        const fieldValues = generateWatchOutput(
+          _name.current as InternalFieldName | InternalFieldName[],
+          control._names,
+          formState.values || control._formValues,
         );
 
         updateValue(
